splits functionality out of uri module
